♻️ Refactor configuration helper and card render
